refactor: Move project configuration to config module
